@@ -1,15 +1,14 @@
 ﻿using System.Collections.Generic;
 using SharpYaml.Serialization;
 
-namespace WebcamSettings
+namespace WebcamSettings;
+
+internal sealed class Camera
 {
-    internal sealed class Camera
-    {
-        [YamlMember(0)]
-        public string DevicePath { get; init; }
+    [YamlMember(0)]
+    public required string DevicePath { get; init; }
 
-        public Dictionary<string, CameraSetting> CameraControls { get; private set; } = new();
+    public Dictionary<string, CameraSetting> CameraControls { get; } = [];
 
-        public Dictionary<string, CameraSetting> VideoControls { get; private set; } = new();
-    }
+    public Dictionary<string, CameraSetting> VideoControls { get; } = [];
 }
