@@ -2,57 +2,59 @@
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
 
-namespace WebcamSettings
+namespace WebcamSettings;
+
+internal sealed class CleanupContainer : IDisposable
 {
-    internal sealed class CleanupContainer : IDisposable
+    private List<object>? _resources = [];
+
+    public void Add(object resource)
     {
-        private List<object> _resources = new();
+        ArgumentNullException.ThrowIfNull(resource, nameof(resource));
+        ObjectDisposedException.ThrowIf(_resources == null, this);
 
-        public void Add(object resource)
+        if (resource is IDisposable || resource is IntPtr || resource.GetType().IsCOMObject)
         {
-            if (resource is null)
-            {
-                throw new ArgumentNullException(nameof(resource));
-            }
+            _resources.Add(resource);
+        }
+    }
 
-            if (_resources == null)
-            {
-                throw new ObjectDisposedException(typeof(CleanupContainer).FullName);
-            }
-
-            if (resource is IDisposable || resource is IntPtr || resource.GetType().IsCOMObject)
-            {
-                _resources.Add(resource);
-            }
+    void IDisposable.Dispose()
+    {
+        if (_resources == null)
+        {
+            return;
         }
 
-        void IDisposable.Dispose()
+        foreach (object resource in _resources)
         {
-            foreach (object resource in _resources)
-            {
-                EnsureCleanup(resource);
-            }
-
-            _resources = null;
+            EnsureCleanup(resource);
         }
 
-        private static void EnsureCleanup(object resource)
+        _resources = null;
+    }
+
+    private static void EnsureCleanup(object resource)
+    {
+        switch (resource)
         {
-            if (resource is IDisposable disposable)
-            {
+            case IDisposable disposable:
                 disposable.Dispose();
-            }
-            else if (OperatingSystem.IsWindows())
-            {
-                if (resource is IntPtr ptr)
+                break;
+            default:
+                if (OperatingSystem.IsWindows())
                 {
-                    Marshal.Release(ptr);
+                    if (resource is nint ptr)
+                    {
+                        Marshal.Release(ptr);
+                    }
+                    else if (resource.GetType().IsCOMObject)
+                    {
+                        while (Marshal.ReleaseComObject(resource) > 0) { }
+                    }
                 }
-                else if (resource.GetType().IsCOMObject)
-                {
-                    while (Marshal.ReleaseComObject(resource) > 0) { }
-                }
-            }
+
+                break;
         }
     }
 }
