@@ -1,16 +1,10 @@
 ﻿using CommandLine;
 
-namespace WebcamSettings
-{
-    [Verb("set", HelpText = "Set camera settings from specified file.")]
-    internal sealed class SetOptions
-    {
-        public SetOptions(string sourceFile)
-        {
-            SourceFile = sourceFile;
-        }
+namespace WebcamSettings;
 
-        [Value(0, MetaName = "source file", Required = true, HelpText = "Source file path to apply camera settings from.")]
-        public string SourceFile { get; }
-    }
+[Verb("set", HelpText = "Set camera settings from specified file.")]
+internal sealed class SetOptions(string sourceFile)
+{
+    [Value(0, MetaName = "source file", Required = true, HelpText = "Source file path to apply camera settings from.")]
+    public string SourceFile { get; } = sourceFile;
 }
