@@ -1,24 +1,16 @@
 ﻿using System;
 using CommandLine;
+using WebcamSettings;
 
-namespace WebcamSettings
+try
 {
-    internal static class Program
-    {
-        public static int Main(string[] args)
-        {
-            try
-            {
-                return Parser.Default.ParseArguments<StoreOptions, SetOptions, int>(args).MapResult(
-                    (StoreOptions store) => SettingsProcessor.Store(store.TargetFile, store.CameraIndex),
-                    (SetOptions set) => SettingsProcessor.Set(set.SourceFile),
-                    _ => 1);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine($"Critical error occurred: '{e.Message}'");
-                return 1;
-            }
-        }
-    }
+    return Parser.Default.ParseArguments<StoreOptions, SetOptions, int>(args).MapResult(
+        (StoreOptions store) => SettingsProcessor.Store(store.TargetFile, store.CameraIndex),
+        (SetOptions set) => SettingsProcessor.Set(set.SourceFile),
+        _ => 1);
+}
+catch (Exception e)
+{
+    Console.WriteLine($"Critical error occurred: '{e.Message}'");
+    return 1;
 }
